交易规则

+++ framework/src/main/java/cn/lili/modules/member/entity/dos/AtmUserRecharge.java
@@ -6,6 +6,8 @@ import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @TableName("atm_user_recharge")
 @ApiModel(value = "会员充值记录")
@@ -16,7 +18,7 @@ public class AtmUserRecharge extends BaseEntity {
     //会员id
     private String userId;
     private int type; //币种类型
-    private int num;//充币数量
+    private BigDecimal num;//充币数量
     private String fileImg;//上传凭证
     @ApiModelProperty(value = "0:待审核 1:审核通过 2:审核失败")
     private int status;//状态
